moved coverage up one level and updated reference to it
