+++ test-module.js
@@ -4,7 +4,7 @@ const Nightmare = require('./xnightmare.js');
 const minimist = require('minimist');
 const config = require('./folio-ui.config.js');
 const helpers = require('./helpers.js');
-const coverageHelper = require('./test/coverage.js');
+const coverageHelper = require('./coverage.js');
 
 const wdToken = 'WD'; // --run value to selectively apply working directory
 const options = minimist(process.argv.slice(2));
